Made audio stream handling safe for long streams

+++ src/audio/AudioStream.h
@@ -9,10 +9,10 @@ public:
 	virtual ~AudioStream() {}
 	virtual std::unique_ptr<AudioStream> clone(bool reset) const = 0;
 	virtual int getSampleRate() const = 0;
-	virtual int getSampleCount() const = 0;
+	virtual int64_t getSampleCount() const = 0;
 	TimeRange getTruncatedRange() const;
-	virtual int getSampleIndex() const = 0;
-	virtual void seek(int sampleIndex) = 0;
+	virtual int64_t getSampleIndex() const = 0;
+	virtual void seek(int64_t sampleIndex) = 0;
 	bool endOfStream() const;
 	virtual float readSample() = 0;
 };
