+++ src/audio/AudioStreamSegment.h
@@ -8,14 +8,14 @@ public:
 	AudioStreamSegment(const AudioStreamSegment& rhs, bool reset);
 	std::unique_ptr<AudioStream> clone(bool reset) const override;
 	int getSampleRate() const override;
-	int getSampleCount() const override;
-	int getSampleIndex() const override;
-	void seek(int sampleIndex) override;
+	int64_t getSampleCount() const override;
+	int64_t getSampleIndex() const override;
+	void seek(int64_t sampleIndex) override;
 	float readSample() override;
 
 private:
 	std::unique_ptr<AudioStream> audioStream;
-	int sampleOffset, sampleCount;
+	int64_t sampleOffset, sampleCount;
 };
 
 std::unique_ptr<AudioStream> createSegment(std::unique_ptr<AudioStream> audioStream, const TimeRange& range);
