+++ src/audio/UnboundedStream.h
@@ -10,13 +10,13 @@ public:
 	UnboundedStream(const UnboundedStream& rhs, bool reset);
 	std::unique_ptr<AudioStream> clone(bool reset) const override;
 	int getSampleRate() const override;
-	int getSampleCount() const override;
-	int getSampleIndex() const override;
-	void seek(int sampleIndex) override;
+	int64_t getSampleCount() const override;
+	int64_t getSampleIndex() const override;
+	void seek(int64_t sampleIndex) override;
 	float readSample() override;
 
 private:
 	std::unique_ptr<AudioStream> innerStream;
-	int sampleIndex;
+	int64_t sampleIndex;
 	boost::optional<float> firstSample, lastSample;
 };
